+++ internal/services/send_receiver_wallets_invite_service.go
@@ -122,12 +122,9 @@ func (s SendReceiverWalletInviteService) SendInvite(ctx context.Context, receive
 			TenantBaseURL:    *currentTenant.BaseURL,
 		}
 
-		registrationLink, err := wdl.GetSignedRegistrationLink(s.sep10SigningPrivateKey)
+		registrationLink, err := s.GetRegistrationLink(ctx, wdl, organization.IsLinkShortenerEnabled)
 		if err != nil {
-			log.Ctx(ctx).Errorf(
-				"error getting signed registration link to receiver wallet ID %s for wallet ID %s and asset ID %s: %s",
-				rwa.ReceiverWallet.ID, wallet.ID, rwa.Asset.ID, err.Error(),
-			)
+			log.Ctx(ctx).Errorf("getting registration link for receiver wallet ID %s: %v", rwa.ReceiverWallet.ID, err)
 			continue
 		}
 
@@ -208,6 +205,29 @@ func (s SendReceiverWalletInviteService) SendInvite(ctx context.Context, receive
 
 		return nil
 	})
+}
+
+func (s SendReceiverWalletInviteService) GetRegistrationLink(ctx context.Context, wdl WalletDeepLink, isLinkShortenerEnabled bool) (string, error) {
+	registrationLink, err := wdl.GetSignedRegistrationLink(s.sep10SigningPrivateKey)
+	if err != nil {
+		return "", fmt.Errorf("getting signed registration link: %w", err)
+	}
+
+	if !isLinkShortenerEnabled {
+		return registrationLink, nil
+	}
+
+	shortCode, err := s.Models.URLShortener.GetOrCreateShortCode(ctx, registrationLink)
+	if err != nil {
+		return "", fmt.Errorf("creating short URL for registration link: %w", err)
+	}
+
+	shortenedRegistrationLink, err := url.JoinPath(wdl.TenantBaseURL, "r", shortCode)
+	if err != nil {
+		return "", fmt.Errorf("building shortened registration link: %w", err)
+	}
+
+	return shortenedRegistrationLink, nil
 }
 
 // resolveReceiverWalletsPendingRegistration returns the receiver wallets pending registration based on the receiverWalletInvitationData.