+++ stellar-multitenant/internal/httphandler/tenants_handler_test.go
@@ -98,7 +98,7 @@ func Test_TenantHandler_Get(t *testing.T) {
 				{
 					"id": %q,
 					"name": %q,
-					"base_url": null,
+					"base_url": %q,
 					"sdp_ui_base_url": null,
 					"status": "TENANT_CREATED",
 					"is_default": false,
@@ -112,7 +112,7 @@ func Test_TenantHandler_Get(t *testing.T) {
 				{
 					"id": %q,
 					"name": %q,
-					"base_url": null,
+					"base_url": %q,
 					"sdp_ui_base_url": null,
 					"status": "TENANT_CREATED",
 					"is_default": false,
@@ -126,7 +126,7 @@ func Test_TenantHandler_Get(t *testing.T) {
 				{
 					"id": %q,
 					"name": %q,
-					"base_url": null,
+					"base_url": %q,
 					"sdp_ui_base_url": null,
 					"status": "TENANT_DEACTIVATED",
 					"is_default": false,
@@ -139,11 +139,11 @@ func Test_TenantHandler_Get(t *testing.T) {
 				}
 			]
 		`,
-			tnt1.ID, tnt1.Name, tnt1.CreatedAt.Format(time.RFC3339Nano), tnt1.UpdatedAt.Format(time.RFC3339Nano),
+			tnt1.ID, tnt1.Name, *tnt1.BaseURL, tnt1.CreatedAt.Format(time.RFC3339Nano), tnt1.UpdatedAt.Format(time.RFC3339Nano),
 			*tnt1.DistributionAccountAddress, schema.DistributionAccountStellarDBVault, schema.AccountStatusActive,
-			tnt2.ID, tnt2.Name, tnt2.CreatedAt.Format(time.RFC3339Nano), tnt2.UpdatedAt.Format(time.RFC3339Nano),
+			tnt2.ID, tnt2.Name, *tnt2.BaseURL, tnt2.CreatedAt.Format(time.RFC3339Nano), tnt2.UpdatedAt.Format(time.RFC3339Nano),
 			*tnt2.DistributionAccountAddress, schema.DistributionAccountStellarDBVault, schema.AccountStatusActive,
-			deactivatedTnt.ID, deactivatedTnt.Name, deactivatedTnt.CreatedAt.Format(time.RFC3339Nano), deactivatedTnt.UpdatedAt.Format(time.RFC3339Nano),
+			deactivatedTnt.ID, deactivatedTnt.Name, *deactivatedTnt.BaseURL, deactivatedTnt.CreatedAt.Format(time.RFC3339Nano), deactivatedTnt.UpdatedAt.Format(time.RFC3339Nano),
 			*deactivatedTnt.DistributionAccountAddress, schema.DistributionAccountStellarDBVault, schema.AccountStatusActive,
 		)
 		assert.JSONEq(t, expectedRespBody, string(respBody))
@@ -168,7 +168,7 @@ func Test_TenantHandler_Get(t *testing.T) {
 			{
 				"id": %q,
 				"name": %q,
-				"base_url": null,
+				"base_url": %q,
 				"sdp_ui_base_url": null,
 				"status": "TENANT_CREATED",
 				"is_default": false,
@@ -179,7 +179,7 @@ func Test_TenantHandler_Get(t *testing.T) {
 				"distribution_account_type": %q,
 				"distribution_account_status": %q
 			}
-		`, tnt1.ID, tnt1.Name, tnt1.CreatedAt.Format(time.RFC3339Nano), tnt1.UpdatedAt.Format(time.RFC3339Nano),
+		`, tnt1.ID, tnt1.Name, *tnt1.BaseURL, tnt1.CreatedAt.Format(time.RFC3339Nano), tnt1.UpdatedAt.Format(time.RFC3339Nano),
 			*tnt1.DistributionAccountAddress, schema.DistributionAccountStellarDBVault, schema.AccountStatusActive,
 		)
 		assert.JSONEq(t, expectedRespBody, string(respBody))
@@ -204,7 +204,7 @@ func Test_TenantHandler_Get(t *testing.T) {
 			{
 				"id": %q,
 				"name": %q,
-				"base_url": null,
+				"base_url": %q,
 				"sdp_ui_base_url": null,
 				"status": "TENANT_CREATED",
 				"is_default": false,
@@ -215,7 +215,7 @@ func Test_TenantHandler_Get(t *testing.T) {
 				"distribution_account_type": %q,
 				"distribution_account_status": %q
 			}
-		`, tnt2.ID, tnt2.Name, tnt2.CreatedAt.Format(time.RFC3339Nano), tnt2.UpdatedAt.Format(time.RFC3339Nano),
+		`, tnt2.ID, tnt2.Name, *tnt2.BaseURL, tnt2.CreatedAt.Format(time.RFC3339Nano), tnt2.UpdatedAt.Format(time.RFC3339Nano),
 			*tnt2.DistributionAccountAddress, schema.DistributionAccountStellarDBVault, schema.AccountStatusActive,
 		)
 		assert.JSONEq(t, expectedRespBody, string(respBody))
@@ -359,6 +359,7 @@ func Test_TenantHandler_Post(t *testing.T) {
 			"receivers",
 			"receivers_audit",
 			"sdp_migrations",
+			"short_urls",
 			"wallets",
 			"wallets_assets",
 		}
@@ -860,7 +861,7 @@ func Test_TenantHandler_Patch_success(t *testing.T) {
 			expectedBodyFn: func(tnt *tenant.Tenant) map[string]interface{} {
 				return map[string]interface{}{
 					"id":                           tnt.ID,
-					"base_url":                     nil,
+					"base_url":                     *tnt.BaseURL,
 					"sdp_ui_base_url":              nil,
 					"status":                       string(tenant.DeactivatedTenantStatus),
 					"distribution_account_address": "GCTNUNQVX7BNIP5AUWW2R4YC7G6R3JGUDNMGT7H62BGBUY4A4V6ROAAH",
@@ -888,7 +889,7 @@ func Test_TenantHandler_Patch_success(t *testing.T) {
 			expectedBodyFn: func(tnt *tenant.Tenant) map[string]interface{} {
 				return map[string]interface{}{
 					"id":                           tnt.ID,
-					"base_url":                     nil,
+					"base_url":                     *tnt.BaseURL,
 					"sdp_ui_base_url":              "http://ui.valid.com",
 					"status":                       string(tenant.ActivatedTenantStatus),
 					"distribution_account_address": "GCTNUNQVX7BNIP5AUWW2R4YC7G6R3JGUDNMGT7H62BGBUY4A4V6ROAAH",